+++ store/src/main/java/com/zimbra/cs/service/mail/ToXML.java
@@ -10,7 +10,6 @@ import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.google.common.io.Closeables;
-import com.zextras.mailbox.smartlinks.SmartLinkUtils;
 import com.zimbra.common.account.Key.AccountBy;
 import com.zimbra.common.calendar.CalendarUtil;
 import com.zimbra.common.calendar.Geo;
@@ -2515,22 +2514,13 @@ public final class ToXML {
       boolean serializeType)
       throws ServiceException {
 
-    if (item instanceof Message && isMimeMessageInvalid((Message) item)) {
-      return parent;
-    }
-
     String name =
         serializeType && item instanceof Chat ? MailConstants.E_CHAT : MailConstants.E_MSG;
     Element elem = parent.addNonUniqueElement(name);
     // DO NOT encode the item-id here, as some Invite-Messages-In-CalendarItems have special
     // item-id's
     if (needToOutput(fields, Change.SIZE)) {
-      if (item instanceof Message) {
-        Message msg = (Message) item;
-        elem.addAttribute(MailConstants.A_SIZE, SmartLinkUtils.getSmartLinkAwareMimeMessageSize(msg));
-      }else{
         elem.addAttribute(MailConstants.A_SIZE, item.getSize());
-      }
     }
 
     if (needToOutput(fields, Change.DATE)) {
@@ -2568,15 +2558,6 @@ public final class ToXML {
       elem.addAttribute(MailConstants.A_IMAP_UID, item.getImapUid());
     }
     return elem;
-  }
-
-  private static boolean isMimeMessageInvalid(Message msg) {
-    try {
-      msg.getMimeMessage();
-      return false;
-    } catch (ServiceException e) {
-      return true;
-    }
   }
 
   private static void encodeTimeZoneMap(Element parent, TimeZoneMap tzmap) {